CLIPCUTTER2/requirements.txt
Flask==2.2.2
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.104.0
ffmpeg-python==0.2.0
python-dotenv==1.0.0
